WAF-396 - Fix and changes requested at code review
